Relocation apps.network package
