+++ bcs-app/backend/kube_core/hpa/views.py
@@ -23,12 +23,12 @@ from backend.accounts import bcs_perm
 from backend.apps.application.base_views import BaseAPI
 from backend.apps.configuration.constants import K8sResourceName
 from backend.apps.constants import ALL_LIMIT
-from backend.apps.network.serializers import BatchResourceSLZ
 from backend.apps.resource.views import ResourceOperate
 from backend.components import paas_cc
 from backend.kube_core.hpa import constants, utils
 from backend.resources.hpa import exceptions as hpa_exceptions
 from backend.resources.hpa import hpa as hpa_client
+from backend.uniapps.network.serializers import BatchResourceSLZ
 from backend.utils.error_codes import error_codes
 from backend.utils.renderers import BKAPIRenderer
 from backend.utils.response import BKAPIResponse
