+++ bcs-app/backend/uniapps/network/utils.py
@@ -25,10 +25,10 @@ from backend.apps.constants import BACKEND_IMAGE_PATH, CONTROLLER_IMAGE_PATH
 from backend.apps.instance import constants as inst_constants
 from backend.apps.instance.funutils import render_mako_context
 from backend.apps.instance.generator import handel_custom_network_mode, handle_intersection_item
-from backend.apps.network.constants import K8S_NGINX_INGRESS_CONTROLLER_CHART_VALUES, MESOS_LB_NAMESPACE
-from backend.apps.network.models import MesosLoadBlance
 from backend.components import paas_cc
 from backend.components.bcs import mesos
+from backend.uniapps.network.constants import K8S_NGINX_INGRESS_CONTROLLER_CHART_VALUES, MESOS_LB_NAMESPACE
+from backend.uniapps.network.models import MesosLoadBlance
 from backend.utils.basic import getitems
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
