+++ bcs-app/backend/uniapps/network/views/charts/releases.py
