+++ bcs-app/backend/uniapps/network/views/charts/versions.py
@@ -15,10 +15,10 @@ from rest_framework import viewsets
 from rest_framework.renderers import BrowsableAPIRenderer
 from rest_framework.response import Response
 
-from backend.apps.network import constants, serializers
-from backend.apps.network.constants import K8S_LB_CHART_NAME, K8S_LB_NAMESPACE
-from backend.apps.network.views.charts.releases import HelmReleaseMixin
 from backend.bcs_k8s.helm.models import ChartVersion
+from backend.uniapps.network import constants, serializers
+from backend.uniapps.network.constants import K8S_LB_CHART_NAME, K8S_LB_NAMESPACE
+from backend.uniapps.network.views.charts.releases import HelmReleaseMixin
 from backend.utils.renderers import BKAPIRenderer
