+++ bcs-app/backend/uniapps/network/views/lb/k8s.py
@@ -27,12 +27,6 @@ from backend.apps.application.utils import APIResponse
 from backend.apps.cluster.constants import DEFAULT_SYSTEM_LABEL_KEYS
 from backend.apps.cluster.models import NodeLabel
 from backend.apps.cluster.serializers import NodeLabelSLZ, NodeLabelUpdateSLZ
-from backend.apps.network import constants, serializers
-from backend.apps.network.constants import K8S_LB_CHART_NAME, K8S_LB_LABEL, K8S_LB_NAMESPACE
-from backend.apps.network.models import K8SLoadBlance
-from backend.apps.network.serializers import NginxIngressSLZ, UpdateK8SLoadBalancerSLZ
-from backend.apps.network.utils import render_helm_values
-from backend.apps.network.views.charts.releases import HelmReleaseMixin
 from backend.bcs_k8s.app.models import App
 from backend.bcs_k8s.app.views import AppViewBase
 from backend.bcs_k8s.helm.models import Chart, ChartVersion
@@ -41,6 +35,12 @@ from backend.components.bcs.k8s import K8SClient
 from backend.resources.cluster import utils as cluster_utils
 from backend.resources.namespace import namespace
 from backend.resources.namespace import utils as ns_utils
+from backend.uniapps.network import constants, serializers
+from backend.uniapps.network.constants import K8S_LB_CHART_NAME, K8S_LB_LABEL, K8S_LB_NAMESPACE
+from backend.uniapps.network.models import K8SLoadBlance
+from backend.uniapps.network.serializers import NginxIngressSLZ, UpdateK8SLoadBalancerSLZ
+from backend.uniapps.network.utils import render_helm_values
+from backend.uniapps.network.views.charts.releases import HelmReleaseMixin
 from backend.utils.error_codes import error_codes
 from backend.utils.renderers import BKAPIRenderer
